requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asgiref @ file:///opt/miniconda3/conda-bld/asgiref_1760342083468/work
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django @ file:///opt/miniconda3/conda-bld/django_1760519132639/work
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
Markdown==3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.5
openai==2.8.1
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_444946bpsc/croot/sqlparse_1734704466504/work
stack-data==0.6.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
webencodings==0.5.1
yarg==0.1.9
